+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Non-breaking changes:
 * Added `WhileUsed` for reference-counted singletons that get garbage-collected when all consumers' `CoroutineScope`s end.
 * Added `conflatedWorker` and `debouncedWorker`` as simple `flowTransformer`s for the suspend-based `derived`/`coAutoRun`.
 * Added `conflatedMap` helper for mapping first and last elements and - whenever possible - intermediate elements.
+* Added simple `ErrorEvents` interface and `withErrorReporting(eventNotifier) { ... }` for easy error handling.
 * `MutableFlow.tryEmit` now returns a `Boolean`.
 
 ## 2.0.4